PIM-3832: introduce small interfaces for different updates and change references
