+++ spec/Pim/Bundle/EnrichBundle/Processor/MassEdit/EditCommonAttributesProcessorSpec.php
@@ -4,6 +4,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Processor\MassEdit;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
+use Akeneo\Component\StorageUtils\Updater\PropertySetterInterface;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Doctrine\ORM\Repository\AttributeRepository;
 use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
@@ -11,7 +12,6 @@ use Pim\Bundle\CatalogBundle\Model\FamilyInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Pim\Bundle\CatalogBundle\Repository\AttributeRepositoryInterface;
 use Pim\Bundle\CatalogBundle\Repository\ProductMassActionRepositoryInterface;
-use Pim\Bundle\CatalogBundle\Updater\ProductFieldUpdaterInterface;
 use Pim\Component\Connector\Model\JobConfigurationInterface;
 use Pim\Component\Connector\Repository\JobConfigurationRepositoryInterface;
 use Prophecy\Argument;
@@ -23,14 +23,14 @@ use Symfony\Component\Validator\ValidatorInterface;
 class EditCommonAttributesProcessorSpec extends ObjectBehavior
 {
     function let(
-        ProductFieldUpdaterInterface $productFieldUpdater,
+        PropertySetterInterface $propertySetter,
         ValidatorInterface $validator,
         ProductMassActionRepositoryInterface $massActionRepository,
         AttributeRepositoryInterface $attributeRepository,
         JobConfigurationRepositoryInterface $jobConfigurationRepo
     ) {
         $this->beConstructedWith(
-            $productFieldUpdater,
+            $propertySetter,
             $validator,
             $massActionRepository,
             $attributeRepository,
@@ -50,7 +50,7 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
 
     function it_sets_values_to_attributes(
         $validator,
-        $productFieldUpdater,
+        $propertySetter,
         FamilyInterface $family,
         AttributeInterface $attribute,
         AttributeRepository $attributeRepository,
@@ -81,14 +81,14 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $product->getFamily()->willReturn($family);
         $stepExecution->incrementSummaryInfo('mass_edited')->shouldBeCalled();
 
-        $productFieldUpdater->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
 
         $this->process($product);
     }
 
     function it_sets_invalid_values_to_attributes(
         $validator,
-        $productFieldUpdater,
+        $propertySetter,
         FamilyInterface $family,
         AttributeInterface $attribute,
         AttributeRepository $attributeRepository,
@@ -120,7 +120,7 @@ class EditCommonAttributesProcessorSpec extends ObjectBehavior
         $family->hasAttribute($attribute)->willReturn(true);
         $product->getFamily()->willReturn($family);
 
-        $productFieldUpdater->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
+        $propertySetter->setData($product, 'categories', ['office', 'bedroom'], [])->shouldBeCalled();
         $this->setStepExecution($stepExecution);
         $stepExecution->addWarning(Argument::cetera())->shouldBeCalled();
         $stepExecution->incrementSummaryInfo('skipped_products')->shouldBeCalled();
