+++ src/Akeneo/Component/StorageUtils/Updater/ObjectUpdaterInterface.php
@@ -3,13 +3,13 @@
 namespace Akeneo\Component\StorageUtils\Updater;
 
 /**
- * Updates an object
+ * Updates an object with a set of data
  *
  * @author    Nicolas Dupont <nicolas@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface UpdaterInterface
+interface ObjectUpdaterInterface
 {
     /**
      * Updates an object (erase the current data)
@@ -20,7 +20,7 @@ interface UpdaterInterface
      *
      * @throws \InvalidArgumentException
      *
-     * @return UpdaterInterface
+     * @return ObjectUpdaterInterface
      */
     public function update($object, array $data, array $options = []);
 }
