+++ src/Pim/Bundle/CatalogBundle/Command/UpdateProductCommand.php
@@ -2,7 +2,10 @@
 
 namespace Pim\Bundle\CatalogBundle\Command;
 
-use Akeneo\Component\StorageUtils\Updater\UpdaterInterface;
+use Akeneo\Component\StorageUtils\Updater\PropertyAdderInterface;
+use Akeneo\Component\StorageUtils\Updater\PropertyCopierInterface;
+use Akeneo\Component\StorageUtils\Updater\PropertyRemoverInterface;
+use Akeneo\Component\StorageUtils\Updater\PropertySetterInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductInterface;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputArgument;
@@ -166,7 +169,7 @@ class UpdateProductCommand extends ContainerAwareCommand
      */
     protected function applySetData(ProductInterface $product, array $update)
     {
-        $updater = $this->getUpdater();
+        $updater = $this->getProductPropertySetter();
         $updater->setData(
             $product,
             $update['field'],
@@ -181,7 +184,7 @@ class UpdateProductCommand extends ContainerAwareCommand
      */
     protected function applyCopyData(ProductInterface $product, array $update)
     {
-        $updater = $this->getUpdater();
+        $updater = $this->getProductPropertyCopier();
         $updater->copyData(
             $product,
             $product,
@@ -202,7 +205,7 @@ class UpdateProductCommand extends ContainerAwareCommand
      */
     protected function applyAddData(ProductInterface $product, array $update)
     {
-        $updater = $this->getUpdater();
+        $updater = $this->getProductPropertyAdder();
         $updater->addData(
             $product,
             $update['field'],
@@ -217,7 +220,7 @@ class UpdateProductCommand extends ContainerAwareCommand
      */
     protected function applyRemoveData(ProductInterface $product, array $update)
     {
-        $updater = $this->getUpdater();
+        $updater = $this->getProductPropertyRemover();
         $updater->removeData(
             $product,
             $update['field'],
@@ -227,11 +230,35 @@ class UpdateProductCommand extends ContainerAwareCommand
     }
 
     /**
-     * @return UpdaterInterface
+     * @return PropertySetterInterface
      */
-    protected function getUpdater()
+    protected function getProductPropertySetter()
     {
-        return $this->getContainer()->get('pim_catalog.updater.product_field');
+        return $this->getContainer()->get('pim_catalog.updater.product_property_setter');
+    }
+
+    /**
+     * @return PropertyAdderInterface
+     */
+    protected function getProductPropertyAdder()
+    {
+        return $this->getContainer()->get('pim_catalog.updater.product_property_adder');
+    }
+
+    /**
+     * @return PropertyRemoverInterface
+     */
+    protected function getProductPropertyRemover()
+    {
+        return $this->getContainer()->get('pim_catalog.updater.product_property_remover');
+    }
+
+    /**
+     * @return PropertyCopierInterface
+     */
+    protected function getProductPropertyCopier()
+    {
+        return $this->getContainer()->get('pim_catalog.updater.product_property_copier');
     }
 
     /**
