+++ src/Pim/Bundle/CatalogBundle/Resources/config/updaters.yml
@@ -1,6 +1,9 @@
 parameters:
     pim_catalog.updater.product.class:                       Pim\Bundle\CatalogBundle\Updater\ProductUpdater
-    pim_catalog.updater.product_field.class:                 Pim\Bundle\CatalogBundle\Updater\ProductFieldUpdater
+    pim_catalog.updater.product_property_setter.class:       Pim\Bundle\CatalogBundle\Updater\ProductPropertySetter
+    pim_catalog.updater.product_property_adder.class:        Pim\Bundle\CatalogBundle\Updater\ProductPropertyAdder
+    pim_catalog.updater.product_property_copier.class:       Pim\Bundle\CatalogBundle\Updater\ProductPropertyCopier
+    pim_catalog.updater.product_property_remover.class:      Pim\Bundle\CatalogBundle\Updater\ProductPropertyRemover
     pim_catalog.updater.product_template.class:              Pim\Bundle\CatalogBundle\Updater\ProductTemplateUpdater
     pim_catalog.updater.attribute_option.class:              Pim\Bundle\CatalogBundle\Updater\AttributeOptionUpdater
 
@@ -55,21 +58,37 @@ services:
     pim_catalog.updater.product:
         class: %pim_catalog.updater.product.class%
         arguments:
-            - @pim_catalog.updater.product_field
+            - @pim_catalog.updater.product_property_setter
+            - @pim_catalog.updater.product_property_copier
 
-    pim_catalog.updater.product_field:
-        class: %pim_catalog.updater.product_field.class%
+    pim_catalog.updater.product_property_setter:
+        class: %pim_catalog.updater.product_property_setter.class%
         arguments:
             - @pim_catalog.repository.attribute
             - @pim_catalog.updater.setter.registry
-            - @pim_catalog.updater.copier.registry
+
+    pim_catalog.updater.product_property_adder:
+        class: %pim_catalog.updater.product_property_adder.class%
+        arguments:
+            - @pim_catalog.repository.attribute
             - @pim_catalog.updater.adder.registry
+
+    pim_catalog.updater.product_property_remover:
+        class: %pim_catalog.updater.product_property_remover.class%
+        arguments:
+            - @pim_catalog.repository.attribute
             - @pim_catalog.updater.remover.registry
+
+    pim_catalog.updater.product_property_copier:
+        class: %pim_catalog.updater.product_property_copier.class%
+        arguments:
+            - @pim_catalog.repository.attribute
+            - @pim_catalog.updater.copier.registry
 
     pim_catalog.updater.product_template:
         class: %pim_catalog.updater.product_template.class%
         arguments:
-            - @pim_catalog.updater.product_field
+            - @pim_catalog.updater.product_property_setter
 
     pim_catalog.updater.attribute_option:
         class: %pim_catalog.updater.attribute_option.class%
