+++ src/Pim/Bundle/CatalogBundle/Updater/ProductTemplateUpdater.php
@@ -2,6 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Updater;
 
+use Akeneo\Component\StorageUtils\Updater\PropertySetterInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
 
 /**
@@ -13,15 +14,15 @@ use Pim\Bundle\CatalogBundle\Model\ProductTemplateInterface;
  */
 class ProductTemplateUpdater implements ProductTemplateUpdaterInterface
 {
-    /** @var ProductFieldUpdaterInterface */
-    protected $productFieldUpdater;
+    /** @var PropertySetterInterface */
+    protected $propertySetter;
 
     /**
-     * @param ProductFieldUpdaterInterface $productFieldUpdater
+     * @param PropertySetterInterface $propertySetter
      */
-    public function __construct(ProductFieldUpdaterInterface $productFieldUpdater)
+    public function __construct(PropertySetterInterface $propertySetter)
     {
-        $this->productFieldUpdater = $productFieldUpdater;
+        $this->propertySetter = $propertySetter;
     }
 
     /**
@@ -45,7 +46,7 @@ class ProductTemplateUpdater implements ProductTemplateUpdaterInterface
     protected function updateProducts(array $products, $attributeCode, $data)
     {
         foreach ($products as $product) {
-            $this->productFieldUpdater->setData(
+            $this->propertySetter->setData(
                 $product,
                 $attributeCode,
                 $data['value'],
